Factored out handler code to proper package * Renamed files properly
