+++ web/docs/env.txt
@@ -1,0 +1,2 @@
+export HORTUS_WEB_IP=192.168.47.242
+export HORTUS_API_IP=192.168.47.242
