+++ web/templates/plantInfo.gohtml
